lib: Remove IArchString interface
